requirements.txt
clip==0.2.0
matplotlib==3.10.9
numpy==2.4.4
opencv_python==4.12.0.88
opencv_python_headless==4.13.0.92
pandas==3.0.2
Pillow==12.2.0
scikit_learn==1.8.0
scipy==1.17.1
streamlit==1.51.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==5.7.0
ultralytics==8.4.6
wandb==0.26.1
yt_dlp==2026.3.17
